requirements.txt
# Core runtime dependencies
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
joblib>=1.2.0
openpyxl>=3.1.0
spacy>=3.5.0
streamlit>=1.29.0
